# swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"LocationDto": {
			"properties": {
				"latitude": {
					"type": "number",
					"format": "double"
				},
				"longitude": {
					"type": "number",
					"format": "double"
				},
				"timeFrom": {
					"type": "string",
					"format": "date-time"
				},
				"timeTo": {
					"type": "string",
					"format": "date-time"
				}
			},
			"required": [
				"latitude",
				"longitude",
				"timeFrom",
				"timeTo"
			],
			"type": "object",
			"additionalProperties": false
		},
		"AddUserTimelineCommand": {
			"properties": {
				"email": {
					"type": "string"
				},
				"testDate": {
					"type": "string",
					"format": "date-time"
				},
				"testType": {
					"type": "string",
					"enum": [
						"COVID1",
						"COVID2"
					]
				},
				"locations": {
					"items": {
						"$ref": "#/definitions/LocationDto"
					},
					"type": "array"
				}
			},
			"required": [
				"email",
				"testDate",
				"testType",
				"locations"
			],
			"type": "object",
			"additionalProperties": false
		},
		"DeleteUserTimelineDataCommand": {
			"properties": {
				"email": {
					"type": "string"
				}
			},
			"required": [
				"email"
			],
			"type": "object",
			"additionalProperties": false
		},
		"LocationWithScoreDto": {
			"properties": {
				"latitude": {
					"type": "number",
					"format": "double"
				},
				"longitude": {
					"type": "number",
					"format": "double"
				},
				"timeFrom": {
					"type": "string",
					"format": "date-time"
				},
				"timeTo": {
					"type": "string",
					"format": "date-time"
				},
				"score": {
					"type": "number",
					"format": "double"
				}
			},
			"required": [
				"latitude",
				"longitude",
				"timeFrom",
				"timeTo",
				"score"
			],
			"type": "object",
			"additionalProperties": false
		},
		"GetLocationsScoreQuery": {
			"properties": {
				"locations": {
					"items": {
						"$ref": "#/definitions/LocationDto"
					},
					"type": "array"
				}
			},
			"required": [
				"locations"
			],
			"type": "object",
			"additionalProperties": false
		}
	},
	"info": {
		"title": "privacy-vs-germs",
		"version": "1.0.0"
	},
	"paths": {
		"/locations/add": {
			"post": {
				"operationId": "Add",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "command",
						"required": true,
						"schema": {
							"$ref": "#/definitions/AddUserTimelineCommand"
						}
					}
				]
			}
		},
		"/locations/delete": {
			"post": {
				"operationId": "Delete",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "command",
						"required": true,
						"schema": {
							"$ref": "#/definitions/DeleteUserTimelineDataCommand"
						}
					}
				]
			}
		},
		"/locations/calculate-scores": {
			"post": {
				"operationId": "Get",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"$ref": "#/definitions/LocationWithScoreDto"
							},
							"type": "array"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "query",
						"required": true,
						"schema": {
							"$ref": "#/definitions/GetLocationsScoreQuery"
						}
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {}
}